+++ NEWS.md
@@ -7,6 +7,7 @@
 # mlr3viz 0.10.0
 
 - Add plot for `LearnerSurvCoxPH`.
+- Add plot for confidence intervals (`mlr3inferr`)
 
 # mlr3viz 0.9.0